+++ apps/ledger-live-desktop/src/main/index.ts
@@ -9,7 +9,6 @@ import {
   getMainWindowAsync,
   loadWindow,
 } from "./window-lifecycle";
-import { getSentryEnabled, setUserId } from "./internal-lifecycle";
 import db from "./db";
 import debounce from "lodash/debounce";
 import sentry from "~/sentry/main";
@@ -78,9 +77,9 @@ app.on("ready", async () => {
   const user: User = (await db.getKey("app", "user")) as User;
   const userId = user?.id;
   if (userId) {
-    setUserId(userId);
     sentry(() => {
-      const value = getSentryEnabled();
+      // const value = getSentryEnabled();
+      const value = undefined;
       if (value === undefined) return settings?.sentryLogs;
       return value;
     }, userId);
