+++ apps/ledger-live-desktop/src/renderer/live-common-setup.ts
@@ -1,21 +1,24 @@
-import "~/live-common-setup-base";
 import "~/live-common-set-supported-currencies";
+import "~/live-common-setup-base";
 import "./families"; // families may set up their own things
 
+import TransportHttp from "@ledgerhq/hw-transport-http";
+import SpeculosHttpTransport, {
+  SpeculosHttpTransportOpts,
+} from "@ledgerhq/hw-transport-node-speculos-http";
 import VaultTransport from "@ledgerhq/hw-transport-vault";
 import TransportWebHID from "@ledgerhq/hw-transport-webhid";
+import { setDeviceMode } from "@ledgerhq/live-common/hw/actions/app";
 import { registerTransportModule } from "@ledgerhq/live-common/hw/index";
 import { retry } from "@ledgerhq/live-common/promise";
+import { getEnv } from "@ledgerhq/live-env";
 import { TraceContext, listen as listenLogs, trace } from "@ledgerhq/logs";
 import { getUserId } from "~/helpers/user";
 import { setEnvOnAllThreads } from "./../helpers/env";
-import { IPCTransport } from "./IPCTransport";
 import logger from "./logger";
-import { currentMode, setDeviceMode } from "@ledgerhq/live-common/hw/actions/app";
 
 setEnvOnAllThreads("USER_ID", getUserId());
 
-const originalDeviceMode = currentMode;
 const vaultTransportPrefixID = "vault-transport:";
 
 // Listens to logs from `@ledgerhq/logs` (happening on the renderer process) and transfers them to the LLD logger system
@@ -25,61 +28,43 @@ listenLogs(({ id, date, ...log }) => {
   logger.debug(log);
 });
 
-// listenLogs(log => {
-//   console.log(log.type + ": " + log.message);
-// });
+if (getEnv("SPECULOS_API_PORT")) {
+  const req: Record<string, number> = {
+    apiPort: getEnv("SPECULOS_API_PORT"),
+  };
 
-// This defines our IPC Transport that will proxy to an internal process (we shouldn't use node-hid on renderer)
-// registerTransportModule({
-//   id: "ipc",
-//   open: (id: string, timeoutMs?: number, context?: TraceContext) => {
-//     // id could be another type of transport such as vault-transport
-//     if (id.startsWith(vaultTransportPrefixID)) return;
-
-//     if (originalDeviceMode !== currentMode) {
-//       setDeviceMode(originalDeviceMode);
-//     }
-
-//     trace({
-//       type: "renderer-setup",
-//       message: "Open called on registered module",
-//       data: {
-//         transport: "IPCTransport",
-//         timeoutMs,
-//       },
-//       context: {
-//         openContext: context,
-//       },
-//     });
-
-//     // Retries in the `renderer` process if the open failed. No retry is done in the `internal` process to avoid multiplying retries.
-//     return retry(() => IPCTransport.open(id, timeoutMs, context), {
-//       interval: 500,
-//       maxRetry: 4,
-//     });
-//   },
-//   disconnect: () => Promise.resolve(),
-// });
-
-registerTransportModule({
-  id: "web-hid",
-  open: (id: string, timeoutMs?: number, context?: TraceContext) => {
-    trace({
-      type: "renderer-setup",
-      message: "Open called on registered module",
-      data: {
-        transport: "TransportWebHID",
-        timeoutMs,
-      },
-      context: {
-        openContext: context,
-      },
-    });
-    return retry(() => TransportWebHID.create(timeoutMs), { interval: 500, maxRetry: 4 });
-    // return TransportWebHID.open();
-  },
-  disconnect: () => Promise.resolve(),
-});
+  registerTransportModule({
+    id: "speculos-http",
+    open: () => retry(() => SpeculosHttpTransport.open(req as SpeculosHttpTransportOpts)),
+    disconnect: () => Promise.resolve(),
+  });
+} else if (getEnv("DEVICE_PROXY_URL")) {
+  const Tr = TransportHttp(getEnv("DEVICE_PROXY_URL").split("|"));
+  registerTransportModule({
+    id: "proxy",
+    open: () => retry(() => Tr.create(3000, 5000)),
+    disconnect: () => Promise.resolve(),
+  });
+} else {
+  registerTransportModule({
+    id: "web-hid",
+    open: (id: string, timeoutMs?: number, context?: TraceContext) => {
+      trace({
+        type: "renderer-setup",
+        message: "Open called on registered module",
+        data: {
+          transport: "TransportWebHID",
+          timeoutMs,
+        },
+        context: {
+          openContext: context,
+        },
+      });
+      return retry(() => TransportWebHID.create(timeoutMs), { interval: 500, maxRetry: 4 });
+    },
+    disconnect: () => Promise.resolve(),
+  });
+}
 
 registerTransportModule({
   id: "vault-transport",
