--- apps/ledger-live-desktop/src/sentry/internal.ts
@@ -1,28 +1,0 @@
-import * as Sentry from "@sentry/node";
-import { Primitive } from "@sentry/types";
-import { init, setShouldSendCallback } from "./install";
-// @ts-expect-error The right type would be typeof SentryMainModule from "@sentry/electron/main"…
-// …but we should avoid importing the whole module here
-const available = init(Sentry, {
-  integrations: [Sentry.httpIntegration()],
-});
-export default (shouldSendCallback: () => boolean, userId: string) => {
-  if (!available) return;
-  setShouldSendCallback(shouldSendCallback);
-  Sentry.setUser({
-    id: userId,
-    ip_address: undefined,
-  });
-};
-export const captureException = (e: Error) => {
-  Sentry.captureException(e);
-};
-export const captureBreadcrumb = (o: Sentry.Breadcrumb) => {
-  Sentry.addBreadcrumb(o);
-};
-export const setTags = (tags: { [key: string]: Primitive }) => {
-  Sentry.setTags(tags);
-};
-export const getSentryIfAvailable = (): typeof Sentry | null => {
-  return available ? Sentry : null;
-};
